requirements.txt
langchain==0.3.27
langchain-openai
langchain-core
langchain-community
docx2txt
pypdf
langchain-chroma
python-multipart
fastapi
uvicorn
chromadb
python-dotenv